Move TestedNumberCard to components/index/cards
